+++ relay-event-normalization/src/normalize/span/description/mod.rs
@@ -3,36 +3,31 @@ mod sql;
 pub use sql::parse_query;
 
 use std::borrow::Cow;
-use std::collections::BTreeMap;
 use std::path::Path;
 
 use itertools::Itertools;
-use relay_event_schema::processor::{self, ProcessingResult};
 use relay_event_schema::protocol::Span;
-use relay_protocol::{Annotated, Remark, RemarkType, Value};
 use url::Url;
 
-use crate::regexes::{REDIS_COMMAND_REGEX, RESOURCE_NORMALIZER_REGEX};
+use crate::regexes::{
+    DB_SQL_TRANSACTION_CORE_DATA_REGEX, REDIS_COMMAND_REGEX, RESOURCE_NORMALIZER_REGEX,
+};
 use crate::span::tag_extraction::HTTP_METHOD_EXTRACTOR_REGEX;
-use crate::transactions::SpanDescriptionRule;
 
 /// Attempts to replace identifiers in the span description with placeholders.
 ///
-/// The resulting scrubbed description is stored in `data.description.scrubbed`, and serves as input
-/// for the span group hash.
-pub(crate) fn scrub_span_description(span: &mut Span, rules: &Vec<SpanDescriptionRule>) {
-    let Some(description) = span.description.as_str() else {
-        return;
-    };
+/// Returns `None` if no scrubbing can be performed.
+pub(crate) fn scrub_span_description(span: &Span) -> Option<String> {
+    let description = span.description.as_str()?;
 
     let db_system = span
         .data
         .value()
         .and_then(|v| v.get("db.system"))
         .and_then(|system| system.as_str());
+    let span_origin = span.origin.as_str();
 
-    let scrubbed = span
-        .op
+    span.op
         .as_str()
         .map(|op| op.split_once('.').unwrap_or((op, "")))
         .and_then(|(op, sub)| match (op, sub) {
@@ -43,9 +38,12 @@ pub(crate) fn scrub_span_description(span: &mut Span, rules: &Vec<SpanDescriptio
                     || sub.contains("mongodb")
                     || sub.contains("redis")
                     || is_legacy_activerecord(sub, db_system)
-                    || is_sql_mongodb(sub, description, db_system)
+                    || is_sql_mongodb(description, db_system)
                 {
                     None
+                // `db.sql.transaction` coming from CoreData need to be scrubbed differently.
+                } else if sub == "sql.transaction" && span_origin == Some("auto.db.core_data") {
+                    scrub_core_data(description)
                 } else {
                     sql::scrub_queries(db_system, description)
                 }
@@ -67,30 +65,28 @@ pub(crate) fn scrub_span_description(span: &mut Span, rules: &Vec<SpanDescriptio
             }
             ("file", _) => scrub_file(description),
             _ => None,
-        });
-
-    if let Some(scrubbed) = scrubbed {
-        span.data
-            .get_or_insert_with(BTreeMap::new)
-            // We don't care what the cause of scrubbing was, since we assume
-            // that after scrubbing the value is sanitized.
-            .insert(
-                "description.scrubbed".to_owned(),
-                Annotated::new(Value::String(scrubbed)),
-            );
-    }
-
-    apply_span_rename_rules(span, rules).ok(); // Only fails on InvalidTransaction
+        })
 }
 
 /// A span declares `op: db.sql.query`, but contains mongodb.
-fn is_sql_mongodb(sub_op: &str, description: &str, db_system: Option<&str>) -> bool {
-    sub_op == "sql.query" && (description.contains("\"$") || db_system == Some("mongodb"))
+fn is_sql_mongodb(description: &str, db_system: Option<&str>) -> bool {
+    description.contains("\"$")
+        || description.contains("({")
+        || description.contains("[{")
+        || description.starts_with('{')
+        || db_system == Some("mongodb")
 }
 
 /// We are unable to parse active record when we do not know which database is being used.
 fn is_legacy_activerecord(sub_op: &str, db_system: Option<&str>) -> bool {
     db_system.is_none() && (sub_op.contains("active_record") || sub_op.contains("activerecord"))
+}
+
+fn scrub_core_data(string: &str) -> Option<String> {
+    match DB_SQL_TRANSACTION_CORE_DATA_REGEX.replace_all(string, "*") {
+        Cow::Owned(scrubbed) => Some(scrubbed),
+        Cow::Borrowed(_) => None,
+    }
 }
 
 fn scrub_http(string: &str) -> Option<String> {
@@ -184,7 +180,12 @@ fn scrub_redis_keys(string: &str) -> Option<String> {
 
 fn scrub_resource_identifiers(mut string: &str) -> Option<String> {
     // Remove query parameters or the fragment.
-    if let Some(pos) = string.find('?') {
+    if string.starts_with("data:") {
+        if let Some(pos) = string.find(';') {
+            return Some(string[..pos].into());
+        }
+        return Some("data:*/*".into());
+    } else if let Some(pos) = string.find('?') {
         string = &string[..pos];
     } else if let Some(pos) = string.find('#') {
         string = &string[..pos];
@@ -193,98 +194,23 @@ fn scrub_resource_identifiers(mut string: &str) -> Option<String> {
         Cow::Owned(scrubbed) => Some(scrubbed),
         Cow::Borrowed(string) => {
             // No IDs scrubbed, but we still want to set something.
-            let url = Url::parse(string).ok()?;
-            let extension = url.path().rsplit_once('.')?.1;
-            let domain = normalize_domain(&url.host()?.to_string(), url.port())?;
-            Some(format!("{domain}/*.{extension}"))
-        }
-    }
-}
-
-/// Applies rules to the span description.
-///
-/// For now, rules are only generated from transaction names, and the
-/// scrubbed value is stored in `span.data[description.scrubbed]` instead of
-/// `span.description` (which remains intact).
-fn apply_span_rename_rules(span: &mut Span, rules: &Vec<SpanDescriptionRule>) -> ProcessingResult {
-    if let Some(op) = span.op.value() {
-        if !op.starts_with("http") {
-            return Ok(());
-        }
-    }
-
-    if rules.is_empty() {
-        return Ok(());
-    }
-
-    // HACK(iker): work-around to scrub the description, in a
-    // context-manager-like approach.
-    //
-    // If data[description.scrubbed] isn't present, we want to scrub
-    // span.description. However, they have different types:
-    // Annotated<Value> vs Annotated<String>. The simplest and fastest
-    // solution I found is to add span.description to span.data if it
-    // doesn't exist already, scrub it, and remove it if we did nothing.
-    let previously_scrubbed = span
-        .data
-        .value()
-        .map(|d| d.get("description.scrubbed"))
-        .is_some();
-    if !previously_scrubbed {
-        if let Some(description) = span.description.clone().value() {
-            span.data
-                .value_mut()
-                .get_or_insert_with(BTreeMap::new)
-                .insert(
-                    "description.scrubbed".to_owned(),
-                    Annotated::new(Value::String(description.to_owned())),
-                );
-        }
-    }
-
-    let mut scrubbed = false;
-
-    if let Some(data) = span.data.value_mut() {
-        if let Some(description) = data.get_mut("description.scrubbed") {
-            processor::apply(description, |name, meta| {
-                if let Value::String(s) = name {
-                    let result = rules.iter().find_map(|rule| {
-                        rule.match_and_apply(Cow::Borrowed(s))
-                            .map(|new_name| (rule.pattern.compiled().pattern(), new_name))
-                    });
-
-                    if let Some((applied_rule, new_name)) = result {
-                        scrubbed = true;
-                        if *s != new_name {
-                            meta.add_remark(Remark::new(
-                                RemarkType::Substituted,
-                                // Setting a different format to not get
-                                // confused by the actual `span.description`.
-                                format!("description.scrubbed:{}", applied_rule),
-                            ));
-                            *name = Value::String(new_name);
-                        }
-                    }
+            // If we managed to parse the URL, we'll try to get an extension.
+            if let Ok(url) = Url::parse(string) {
+                let domain = normalize_domain(&url.host()?.to_string(), url.port())?;
+                // If there is an extension, we add it to the domain.
+                if let Some(extension) = url.path().rsplit_once('.') {
+                    return Some(format!("{domain}/*.{}", extension.1));
                 }
-
-                Ok(())
-            })?;
+                return Some(format!("{domain}/*"));
+            }
+            Some(string.into())
         }
     }
-
-    if !previously_scrubbed && !scrubbed {
-        span.data
-            .value_mut()
-            .as_mut()
-            .and_then(|data| data.remove("description.scrubbed"));
-    }
-
-    Ok(())
 }
 
 #[cfg(test)]
 mod tests {
-    use relay_protocol::get_value;
+    use relay_protocol::Annotated;
     use similar_asserts::assert_eq;
 
     use super::*;
@@ -294,7 +220,7 @@ mod tests {
 
         // Same output and input means the input was already scrubbed.
         // An empty output `""` means the input wasn't scrubbed and Relay didn't scrub it.
-        ($name:ident, $description_in:literal, $op_in:literal, $output:literal) => {
+        ($name:ident, $description_in:literal, $op_in:literal, $expected:literal) => {
             #[test]
             fn $name() {
                 let json = format!(
@@ -318,23 +244,12 @@ mod tests {
                     .description
                     .set_value(Some($description_in.into()));
 
-                scrub_span_description(span.value_mut().as_mut().unwrap(), &vec![]);
+                let scrubbed = scrub_span_description(span.value_mut().as_mut().unwrap());
 
-                if $output == "" {
-                    assert!(span
-                        .value()
-                        .and_then(|span| span.data.value())
-                        .and_then(|data| data.get("description.scrubbed"))
-                        .is_none());
+                if $expected == "" {
+                    assert!(scrubbed.is_none());
                 } else {
-                    assert_eq!(
-                        $output,
-                        span.value()
-                            .and_then(|span| span.data.value())
-                            .and_then(|data| data.get("description.scrubbed"))
-                            .and_then(|an_value| an_value.as_str())
-                            .unwrap()
-                    );
+                    assert_eq!(scrubbed.unwrap(), $expected);
                 }
             }
         };
@@ -547,7 +462,7 @@ mod tests {
     );
 
     span_description_test!(
-        span_description_fil_no_extension,
+        span_description_file_with_no_extension,
         "somefilenamewithnoextension",
         "file.read",
         "*"
@@ -559,6 +474,43 @@ mod tests {
         "resource.img",
         "https://data.domain.com/data/guide*.gif"
     );
+
+    span_description_test!(
+        resource_script_with_no_extension,
+        "https://www.domain.com/page?id=1234567890",
+        "resource.script",
+        "*.domain.com/*"
+    );
+
+    span_description_test!(
+        resource_script_with_no_domain,
+        "/page.js?action=name",
+        "resource.script",
+        "/page.js"
+    );
+
+    span_description_test!(
+        resource_script_with_no_domain_no_extension,
+        "/page?action=name",
+        "resource.script",
+        "/page"
+    );
+
+    span_description_test!(
+        resource_img_embedded,
+        "data:image/svg+xml;base64,PHN2ZyB4bW",
+        "resource.img",
+        "data:image/svg+xml"
+    );
+
+    span_description_test!(
+        db_category_with_mongodb_query,
+        "find({some_id:1234567890},{limit:100})",
+        "db",
+        ""
+    );
+
+    span_description_test!(db_category_with_not_sql, "{someField:someValue}", "db", "");
 
     #[test]
     fn informed_sql_parser() {
@@ -576,14 +528,8 @@ mod tests {
 
         let mut span = Annotated::<Span>::from_json(json).unwrap();
         let span = span.value_mut().as_mut().unwrap();
-        scrub_span_description(span, &vec![]);
-        let scrubbed = span
-            .data
-            .value()
-            .and_then(|d| d.get("description.scrubbed"))
-            .and_then(|v| v.value())
-            .and_then(|v| v.as_str());
-        assert_eq!(scrubbed, Some("SELECT %s"));
+        let scrubbed = scrub_span_description(span);
+        assert_eq!(scrubbed.as_deref(), Some("SELECT %s"));
     }
 
     #[test]
@@ -595,8 +541,7 @@ mod tests {
 
         let mut span = Annotated::<Span>::from_json(json).unwrap();
 
-        scrub_span_description(span.value_mut().as_mut().unwrap(), &vec![]);
-        let scrubbed = get_value!(span.data["description.scrubbed"]);
+        let scrubbed = scrub_span_description(span.value_mut().as_mut().unwrap());
 
         // When db.system is missing, no scrubbed description (i.e. no group) is set.
         assert!(scrubbed.is_none());
@@ -614,10 +559,42 @@ mod tests {
 
         let mut span = Annotated::<Span>::from_json(json).unwrap();
 
-        scrub_span_description(span.value_mut().as_mut().unwrap(), &vec![]);
-        let scrubbed = get_value!(span.data["description.scrubbed"]!);
+        let scrubbed = scrub_span_description(span.value_mut().as_mut().unwrap());
 
         // Can be scrubbed with db system.
-        assert_eq!(scrubbed.as_str(), Some("SELECT a FROM b"));
+        assert_eq!(scrubbed.as_deref(), Some("SELECT a FROM b"));
+    }
+
+    #[test]
+    fn core_data() {
+        let json = r#"{
+            "description": "INSERTED 1 'UAEventData'",
+            "op": "db.sql.transaction",
+            "origin": "auto.db.core_data"
+        }"#;
+
+        let mut span = Annotated::<Span>::from_json(json).unwrap();
+
+        let scrubbed = scrub_span_description(span.value_mut().as_mut().unwrap());
+
+        assert_eq!(scrubbed.as_deref(), Some("INSERTED * 'UAEventData'"));
+    }
+
+    #[test]
+    fn multiple_core_data() {
+        let json = r#"{
+            "description": "UPDATED 1 'QueuedRequest', DELETED 1 'QueuedRequest'",
+            "op": "db.sql.transaction",
+            "origin": "auto.db.core_data"
+        }"#;
+
+        let mut span = Annotated::<Span>::from_json(json).unwrap();
+
+        let scrubbed = scrub_span_description(span.value_mut().as_mut().unwrap());
+
+        assert_eq!(
+            scrubbed.as_deref(),
+            Some("UPDATED * 'QueuedRequest', DELETED * 'QueuedRequest'")
+        );
     }
 }